data/models/own/3D-Segmentation/src/requirements.txt
torchio==0.19.7
torchvision==0.17.2+cu118
pytorch-lightning==2.2.4